requirements.txt
Django==1.11
psycopg2==2.7.1
djangorestframework==3.6.2
django-compressor==2.1.1
django-bootstrap3==8.2.3
django-wysiwyg-redactor==0.5.1
Pillow==4.1.1

ipdb==0.10.3
ipython==5.3.0
django-extensions==1.7.9

gunicorn==19.7.1
whitenoise==3.3.0
psycopg2==2.7.1
dj-database-url==0.4.2
